+++ webextension/js/panel/panel-init.js
@@ -75,7 +75,6 @@ window.onload = function () {
 
 		await import('../../lib/throttle.js');
 		await import('../panel/tabMover.js');
-		await import('../variousFeatures/website-data.js');
 		await import('../panel/panel.js');
 	})()
 		.catch(console.error)